fix: query and schema todos

+++ src/service/TodoService.js
@@ -1,44 +1,59 @@
-import { NotFoundError } from "../exception/error.js"
-import { CreateSchema, UpadteSchema, FindSchema } from "../validator/TodoSchema.js"
+import { NotFoundError, UnauthorizedError } from "../exception/error.js"
+import Atoi from "../utils/strconv.js"
+import { CreateSchema, FindSchema, UpdateSchema } from "../validator/TodoSchema.js"
 
 const TodoService = (TodoRepository, validator) => ({
-    FindAll: async () => {
-        return await TodoRepository.FindAll()
+    FindAll: async (todo) => {
+        return await TodoRepository.FindAllOwnUser(todo.user.id)
     },
-
-    FindOne: async (id) => {
-        const isValid = validator.validate({ id: id }, FindSchema)
-        validator.check(isValid)
-        const find = await TodoRepository.FindOne(id)
+    FindOne: async (todo) => {
+        const id = Atoi(todo)
+        const user_id = todo.user.id
+        const idValidate = validator.validate({ id: id }, FindSchema)
+        validator.check(idValidate)
+        const user_idValidate = validator.validate({ id: user_id }, FindSchema)
+        validator.check(user_idValidate)
+        const find = await TodoRepository.FindOneOwnUser(id, user_id)
         if (!find) throw NotFoundError()
         return find
     },
 
     Create: async (todo) => {
-        const isValid = validator.validate(todo, CreateSchema)
+        todo.body.user_id = todo.user.id
+        const isValid = validator.validate(todo.body, CreateSchema)
         validator.check(isValid)
-        todo.date = new Date().toISOString()
-        return await TodoRepository.Create(todo)
+        console.log('service create todo', todo.body)
+        return await TodoRepository.Create(todo.body)
     },
 
-    Update: async (id, todo) => {
-        todo.id = id
-        const isValid = validator.validate(todo, UpadteSchema)
+    Update: async (todo) => {
+        console.log('todo body update', todo.params)
+        const id = Atoi(todo)
+        todo.body.id = id
+
+        const isValid = validator.validate(todo.body, UpdateSchema)
         validator.check(isValid)
+
         const find = await TodoRepository.FindOne(id)
+        console.log('ini find', find)
         if (!find) throw NotFoundError()
-        todo.name ? todo.name : todo.name = find.name
-        todo.status ? todo.status : todo.status = find.status
-        todo.date = find.date
-        const result = await TodoRepository.Update(find.id, todo)
+        if (find.user_id !== todo.user.id) throw UnauthorizedError(`you don't have permission to update this todo`)
+
+        todo.body.name ? todo.body.name : todo.body.name = find.name
+        todo.body.status ? todo.body.status : todo.body.status = find.status
+
+        console.log('todo body update after reasign', todo.body)
+        const result = await TodoRepository.Update(find.id, todo.body)
         return result
     },
 
-    Delete: async (id) => {
+    Delete: async (todo) => {
+        const id = Atoi(todo)
         const isValid = validator.validate({ id: id }, FindSchema)
         validator.check(isValid)
         const find = await TodoRepository.FindOne(id)
         if (!find) throw NotFoundError()
+        if (find.user_id !== todo.user.id) throw UnauthorizedError(`you don't have permission to update this todo`)
         return await TodoRepository.Delete(id)
     }
 })
